+++ FileSystemNTFS/Controller/FileSystemController/MoveFile.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using FileSystemNTFS.BL.Models;
+using MultiuserProtection.Domain;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -15,18 +17,31 @@ namespace FileSystemNTFS.BL.Controller.FileSystemController
 
             if (!File.Exists(oldFullPath) || mftItem == null)
                 return false;
-            DeleteMFTDataFromDir(mftItem);
-            var newFullPath = Path.Combine(newDirPath, fileName + ".bin");
-            File.Move(oldFullPath,newFullPath);
 
-            mftItem.Attributes.FullPath = newFullPath;
-            mftItem.Attributes.ParentsDirectory = mftItem.Attributes.GetParentsDir(newFullPath);
-            AddMFTDataToDir(mftItem);
-            FileSystem.MFTController.Update(mftItem);
+            if (FileSystem.UserController.CurrentUser.AccountType == AccountType.Administrator ||
+                            mftItem.Attributes.AccessFlags.O == AttributeFlags.Modify ||
+                            mftItem.Attributes.AccessFlags.O == AttributeFlags.FullControl ||
+                            (mftItem.Attributes.OwnerId == FileSystem.UserController.CurrentUser.Id && (mftItem.Attributes.AccessFlags.U == AttributeFlags.FullControl || mftItem.Attributes.AccessFlags.U == AttributeFlags.Modify)) ||
+                            (mftItem.Attributes.Groups.Any(FileSystem.UserController.CurrentUser.Groups.Contains) && (mftItem.Attributes.AccessFlags.G == AttributeFlags.FullControl || mftItem.Attributes.AccessFlags.G == AttributeFlags.Modify)))
+            {
+                DeleteMFTDataFromDir(mftItem);
+                var newFullPath = Path.Combine(newDirPath, fileName + ".bin");
+                File.Move(oldFullPath, newFullPath);
+                mftItem.Attributes.FullPath = newFullPath;
+                mftItem.Attributes.ParentsDirectory = mftItem.Attributes.GetParentsDir(newFullPath);
+                AddMFTDataToDir(mftItem);
+                FileSystem.MFTController.Update(mftItem);
+                Save();
 
-            Save();
-
-            return true;
+                return true;
+            }
+            else
+            {
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine("Недостаточно прав!");
+                Console.ResetColor();
+                return false;
+            }
         }
     }
 }
